refactor(nvim-ginkgo): rename cmd module to spec
Rename `cmd.lua` to `spec.lua` to better reflect its purpose of building test specifications. Update all references in the init and test files to use the new module name.

- Rename `cmd.lua` to `spec.lua`
- Update `init.lua` to require `spec` instead of `cmd`
- Update test file references from `cmd` to `spec`
- Modify test descriptions to match new module name
