+++ lua/nvim-ginkgo/init.lua
@@ -4,7 +4,7 @@ local dap = require("nvim-ginkgo.strategies.dap")
 local lib = require("neotest.lib")
 local logger = require("neotest.logging")
 local tree = require("nvim-ginkgo.tree")
-local cmd = require("nvim-ginkgo.cmd")
+local spec = require("nvim-ginkgo.spec")
 local report = require("nvim-ginkgo.report")
 
 ---@class nvim-ginkgo.Adapter: neotest.Adapter
@@ -59,7 +59,7 @@ end
 ---@param args neotest.RunArgs
 ---@return nil | neotest.RunSpec | neotest.RunSpec[]
 function adapter.build_spec(args)
-	return cmd.build(args)
+	return spec.build(args)
 end
 
 ---@async
